Put simplified types in a dedicated place
